third_party/WebKit/LayoutTests/fast/canvas/currentTransform-null.html

<!DOCTYPE html>
<canvas id="canvas" width="100" height="100"></canvas>
<script src="../../resources/js-test.js"></script>
<script>
description('Check that assigning |null| and |undefined| to currentTransform ' +
            'do not cause crashes, but do throw exceptions.');

var canvas = document.getElementById('canvas');
var context = canvas.getContext('2d');
var undef;
shouldThrow("context.currentTransform = null;");
shouldThrow("context.currentTransform = undef;");
</script>
